+++ package.json
@@ -4,11 +4,14 @@
   "license": "MIT",
   "version": "0.3.3",
   "private": true,
+  "workspaces": [
+    "packages/*"
+  ],
   "scripts": {
-    "preinstall": "npx -y only-allow pnpm",
+    "preinstall": "only-allow pnpm",
     "lint": "eslint .",
     "build": "pnpm -r run build",
-    "package": "pnpm build && node ./scripts/prepackage.js && pnpm --filter prettify-ts run package"
+    "package": "pnpm build && pnpm --filter prettify-ts run package && node ./scripts/post-package.js && pnpm install"
   },
   "devDependencies": {
     "@eslint/js": "^9.27.0",
@@ -22,5 +25,8 @@
     "only-allow": "^1.2.1",
     "typescript": "^5.4.5",
     "typescript-eslint": "^8.30.1"
+  },
+  "dependencies": {
+    "@prettify-ts/typescript-plugin": "workspace:*"
   }
 }
