+++ pnpm-workspace.yaml
@@ -1,5 +1,5 @@
 packages:
-  - packages/*
+  - 'packages/*'
 onlyBuiltDependencies:
   - '@vscode/vsce-sign'
   - esbuild
